pyproject.toml
[project]
name = "battleship"
version = "0.2.0"
description = "Battleship"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "fastsession>=0.3.0",
    "numpy>=2.4.2",
    "python-socketio>=5.16.1",
    "uvicorn[standard]>=0.42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "isort>=8.0.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
]
